+++ .storybook/webpack.config.js
@@ -8,8 +8,6 @@ module.exports = ({ config, mode }) => {
         loader: require.resolve('babel-loader'),
         options: {
           presets: [
-            require.resolve('@babel/preset-typescript'),
-            require.resolve('@babel/preset-react'),
             [
               require.resolve('@babel/preset-env'),
               {
@@ -20,7 +18,9 @@ module.exports = ({ config, mode }) => {
                   ]
                 }
               }
-            ]
+            ],
+            require.resolve('@babel/preset-react'),
+            require.resolve('@babel/preset-typescript')
           ],
           plugins: [
             require.resolve('@babel/plugin-proposal-class-properties'),
